requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.17.1
gunicorn==25.3.0
idna==3.11
packaging==26.1
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
uuid6==2025.0.1
